pyproject.toml
[tool.poetry]
name = "llm-math-education"
version = "0.4.0"
description = "Retrieval-backed LLMs for math education"
authors = [
    "Zachary Levonian <zach@digitalharbor.org>"
]
license = "MIT"
readme = "README.md"
packages = [{include = "llm_math_education", from = "src"}]
repository = "https://github.com/levon003/llm-math-education.git"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
poetry = "1.4.0"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
streamlit = "^1.23.1"
pandas = "^2.0.2"
tiktoken = "^0.4.0"
scipy = "^1.11.0"
argon2-cffi = "^21.3.0"

[tool.poetry.group.dev.dependencies]
httpx = "^0.24.1"
jupyter = "^1.0.0"
matplotlib = "^3.7.1"
urllib3 = "<=2.0"
openpyxl = "^3.1.2"
scikit-learn = "^1.2.2"
beautifulsoup4 = "^4.12.2"
black = "^22.12.0"
isort = "^5.12"
flake8 = "^6.0.0"
nbqa = "^1.6.0"
pre-commit = "^2.21.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
jupyterlab = "^4.0.2"
pytest-dotenv = "^0.5.2"
torch = "^2.0.1"
outlines = "^0.0.8"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
supabase = "^1.0.4"

[tool.poetry.group.anserini]
optional = true

[tool.poetry.group.anserini.dependencies]
lightgbm = "^3.3.5"
nmslib = "^2.1.1"
pyserini = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
      .eggs         # exclude a few common directories in the
    | .git          # root of the project
    | .github
    | .gitignore
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | venv
    | _build
    | buck-out
    | build
    | ci
    | data
    | dist
    | docs
    | docsrc
  )/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
virtual_env = "venv"
